fix(bom): 修复 QA 验收发现的缺陷与风险项
- BUG-01: productionQty 校验增加 .int() 约束，禁止小数件数
- BUG-02: 批量导入按钮在 AI 数据加载中也禁用
- RISK-01: useAddBomItem/useDeleteBomItem/useUpdateBomItem 成功后
           invalidate costBreakdown 缓存，确保成本数据自动刷新
- RISK-05: 向导 submitting 状态时禁止关闭 Modal

+++ services/api/src/modules/bom/bom.controller.ts
@@ -23,7 +23,7 @@ const CreateBomSchema = z.object({
 });
 
 const CalcRequirementsSchema = z.object({
-  productionQty: z.coerce.number().positive().max(1_000_000),
+  productionQty: z.coerce.number().int().positive().max(1_000_000),
 });
 
 const AddBomItemSchema = z.object({
